refactor: update all external imports after network package reorganisation
