+++ app/src/org/commcare/network/CommcareRequestGenerator.java
@@ -9,8 +9,8 @@ import com.google.common.collect.Multimap;
 import org.commcare.CommCareApplication;
 import org.commcare.android.database.user.models.ACase;
 import org.commcare.connect.PersonalIdManager;
-import org.commcare.connect.network.ConnectSsoHelper;
-import org.commcare.connect.network.TokenUnavailableException;
+import org.commcare.connect.network.connectId.ConnectSsoHelper;
+import org.commcare.connect.network.connectId.TokenUnavailableException;
 import org.commcare.core.network.AuthInfo;
 import org.commcare.core.network.HTTPMethod;
 import org.commcare.core.network.ModernHttpRequester;
